refactor: move brand-template to @rule-io/client + utils to @rule-io/core
Completes the rcml-cleanup arc. @rule-io/rcml now ships only its canonical
email/ surface (schema, element factories, Rcml* types, validators, RFM/XML
converters); its top-level index.ts is a single `export * from './email/index.js'`.

Two files relocated:
- brand-template.ts (createBrandTemplate, createBrandText, createContentSection,
  resolvePreferredBrandStyle, validateCustomFields, etc.) → @rule-io/client.
  Half of it is the brand-style HTTP bridge, which naturally co-locates with
  the HTTP client; the pure-builder half rides along (splitting is an
  independent follow-up).
- utils.ts (escapeHtml, sanitizeUrl, formatDateForRule) → @rule-io/core.
  Dependency-free primitives used by multiple packages — belong at the root
  of the graph.

Consumers (vendor-shopify, vendor-bookzen, vendor-samfora, cli) split their
`@rule-io/rcml` imports accordingly. createTextNode collision between rcml's
email/create-rcml-element.ts and brand-template is resolved by having the
latter re-export the former so sdk's `export *` sees one canonical symbol.

Vendor packages gained `@rule-io/client` in their package.json dependencies.
Test helpers that previously lived in packages/rcml/tests/helpers.ts moved to
packages/client/tests/helpers.ts.

packages/rcml/src top-level shrinks to 14 LoC (was 1436, -99%).
