+++ packages/cli/src/commands/deploy-bookzen.ts
@@ -11,7 +11,8 @@
 
 import type { Command } from 'commander';
 import { RuleClient } from '@rule-io/client';
-import { resolvePreferredBrandStyle, type RcmlDocument } from '@rule-io/rcml';
+import { type RcmlDocument } from '@rule-io/rcml';
+import { resolvePreferredBrandStyle } from '@rule-io/client';
 import type { CustomFieldMap, VendorConsumerConfig } from '@rule-io/core';
 import { bookzenPreset, BOOKZEN_FIELDS, BOOKZEN_TAGS } from '@rule-io/vendor-bookzen';
 import { createClient } from '../shared/client.js';
