+++ packages/client/src/index.ts
@@ -11,3 +11,4 @@ export { RuleClient } from './client.js';
 export { RULE_API_V2_BASE_URL, RULE_API_V3_BASE_URL, RuleTags } from './constants.js';
 export type { RuleTag } from './constants.js';
 export type * from './types.js';
+export * from './brand-template.js';
