+++ packages/vendor-bookzen/package.json
@@ -1,7 +1,7 @@
 {
   "name": "@rule-io/vendor-bookzen",
   "version": "0.3.0",
-  "description": "Bookzen preset for the Rule.io SDK — hospitality automation flows (reservation confirmation, cancellation, reminder, feedback, request).",
+  "description": "Bookzen preset for the Rule.io SDK \u2014 hospitality automation flows (reservation confirmation, cancellation, reminder, feedback, request).",
   "type": "module",
   "main": "./src/index.js",
   "types": "./src/index.d.ts",
@@ -16,7 +16,14 @@
     "README.md",
     "CHANGELOG.md"
   ],
-  "keywords": ["rule.io", "bookzen", "hospitality", "reservations", "preset", "sdk"],
+  "keywords": [
+    "rule.io",
+    "bookzen",
+    "hospitality",
+    "reservations",
+    "preset",
+    "sdk"
+  ],
   "license": "MIT",
   "repository": {
     "type": "git",
@@ -24,6 +31,7 @@
     "directory": "packages/vendor-bookzen"
   },
   "dependencies": {
+    "@rule-io/client": "0.3.0",
     "@rule-io/core": "0.3.0",
     "@rule-io/rcml": "0.3.0"
   },
