+++ packages/vendor-shopify/tests/ecommerce-templates.test.ts
@@ -8,7 +8,8 @@
  */
 
 import { describe, expect, it } from 'vitest'
-import type { CustomFieldMap } from '@rule-io/core';import { createStatusTrackerSection } from '@rule-io/rcml'
+import type { CustomFieldMap } from '@rule-io/core';
+import { createStatusTrackerSection } from '@rule-io/client';
 import { RuleConfigError } from '@rule-io/core'
 import {
   createAbandonedCartEmail,
